+++ ElvUI_LuckyoneUI/Core/Install.lua
@@ -79,10 +79,10 @@ if L1UI.Retail then
 				PluginInstallFrame.Desc1:SetText('Choose between DBM and BigWigs.')
 				PluginInstallFrame.Desc2:SetText('Importance: |cff4beb2cOptional|r')
 				PluginInstallFrame.Option1:Show()
-				PluginInstallFrame.Option1:SetScript('OnClick', function() L1UI:AddonSetupBW() end)
+				PluginInstallFrame.Option1:SetScript('OnClick', function() L1UI:Setup_BigWigs() end)
 				PluginInstallFrame.Option1:SetText('BigWigs')
 				PluginInstallFrame.Option2:Show()
-				PluginInstallFrame.Option2:SetScript('OnClick', function() L1UI:AddonSetupDBM() end)
+				PluginInstallFrame.Option2:SetScript('OnClick', function() L1UI:Setup_DBM() end)
 				PluginInstallFrame.Option2:SetText('DBM')
 			end,
 			[8] = function()
@@ -90,23 +90,31 @@ if L1UI.Retail then
 				PluginInstallFrame.Desc1:SetText('Please click the button below to apply Luckyones profile for Details! Damage Meter.')
 				PluginInstallFrame.Desc2:SetText('Importance: |cff4beb2cOptional|r')
 				PluginInstallFrame.Option1:Show()
-				PluginInstallFrame.Option1:SetScript('OnClick', function() L1UI:AddonSetupDT() end)
+				PluginInstallFrame.Option1:SetScript('OnClick', function() L1UI:Setup_Details() end)
 				PluginInstallFrame.Option1:SetText('Setup Details')
 			end,
 			[9] = function()
+				PluginInstallFrame.SubTitle:SetFormattedText('Gnosis Profile')
+				PluginInstallFrame.Desc1:SetText('Please click the button below to apply Luckyones profile for Gnosis.\nThis will disable ElvUI castbars.')
+				PluginInstallFrame.Desc2:SetText('Importance: |cff4beb2cOptional|r')
+				PluginInstallFrame.Option1:Show()
+				PluginInstallFrame.Option1:SetScript('OnClick', function() L1UI:Setup_Gnosis() end)
+				PluginInstallFrame.Option1:SetText('Setup Gnosis')
+			end,
+			[10] = function()
 				PluginInstallFrame.SubTitle:SetFormattedText('OmniCD Profile')
 				PluginInstallFrame.Desc1:SetText('Please click the button below to apply Luckyones profile for OmniCD Party CDs.')
 				PluginInstallFrame.Desc2:SetText('Importance: |cff4beb2cOptional|r')
 				PluginInstallFrame.Option1:Show()
-				PluginInstallFrame.Option1:SetScript('OnClick', function() L1UI:AddonSetupOCD() end)
+				PluginInstallFrame.Option1:SetScript('OnClick', function() L1UI:Setup_OmniCD() end)
 				PluginInstallFrame.Option1:SetText('Setup OmniCD')
 			end,
-			[10] = function()
+			[11] = function()
 				PluginInstallFrame.SubTitle:SetFormattedText('WeakAuras Information')
 				PluginInstallFrame.Desc1:SetText('You can find WeakAuras for all 12 Classes and some Utility WeakAuras in the LuckyoneUI config.')
 				PluginInstallFrame.Desc2:SetText('\nElvUI Options -> LuckyoneUI -> WeakAuras')
 			end,
-			[11] = function()
+			[12] = function()
 				PluginInstallFrame.SubTitle:SetText('Installation Complete')
 				PluginInstallFrame.Desc1:SetText('You have completed the installation process, please click "Finished" to reload the UI.')
 				PluginInstallFrame.Desc2:SetText('Importance: |cff4beb2cHigh|r')
@@ -124,9 +132,10 @@ if L1UI.Retail then
 			[6] = 'NamePlates',
 			[7] = 'BossMods',
 			[8] = 'Details',
-			[9] = 'OmniCD',
-			[10] = 'WeakAuras',
-			[11] = 'Installation Complete',
+			[9] = 'Gnosis',
+			[10] = 'OmniCD',
+			[11] = 'WeakAuras',
+			[12] = 'Installation Complete',
 		},
 		StepTitlesColor = {1, 1, 1},
 		StepTitlesColorSelected = {0, 179/255, 1},
@@ -205,10 +214,10 @@ elseif L1UI.Classic then
 				PluginInstallFrame.Desc1:SetText('Choose between DBM and BigWigs.')
 				PluginInstallFrame.Desc2:SetText('Importance: |cff4beb2cOptional|r')
 				PluginInstallFrame.Option1:Show()
-				PluginInstallFrame.Option1:SetScript('OnClick', function() L1UI:AddonSetupBW() end)
+				PluginInstallFrame.Option1:SetScript('OnClick', function() L1UI:Setup_BigWigs() end)
 				PluginInstallFrame.Option1:SetText('BigWigs')
 				PluginInstallFrame.Option2:Show()
-				PluginInstallFrame.Option2:SetScript('OnClick', function() L1UI:AddonSetupDBM() end)
+				PluginInstallFrame.Option2:SetScript('OnClick', function() L1UI:Setup_DBM() end)
 				PluginInstallFrame.Option2:SetText('DBM')
 			end,
 			[8] = function()
@@ -216,7 +225,7 @@ elseif L1UI.Classic then
 				PluginInstallFrame.Desc1:SetText('Please click the button below to apply Luckyones profile for Details! Damage Meter.')
 				PluginInstallFrame.Desc2:SetText('Importance: |cff4beb2cOptional|r')
 				PluginInstallFrame.Option1:Show()
-				PluginInstallFrame.Option1:SetScript('OnClick', function() L1UI:AddonSetupDT() end)
+				PluginInstallFrame.Option1:SetScript('OnClick', function() L1UI:Setup_Details() end)
 				PluginInstallFrame.Option1:SetText('Setup Details')
 			end,
 			[9] = function()
@@ -316,10 +325,10 @@ elseif L1UI.TBC then
 				PluginInstallFrame.Desc1:SetText('Choose between DBM and BigWigs.')
 				PluginInstallFrame.Desc2:SetText('Importance: |cff4beb2cOptional|r')
 				PluginInstallFrame.Option1:Show()
-				PluginInstallFrame.Option1:SetScript('OnClick', function() L1UI:AddonSetupBW() end)
+				PluginInstallFrame.Option1:SetScript('OnClick', function() L1UI:Setup_BigWigs() end)
 				PluginInstallFrame.Option1:SetText('BigWigs')
 				PluginInstallFrame.Option2:Show()
-				PluginInstallFrame.Option2:SetScript('OnClick', function() L1UI:AddonSetupDBM() end)
+				PluginInstallFrame.Option2:SetScript('OnClick', function() L1UI:Setup_DBM() end)
 				PluginInstallFrame.Option2:SetText('DBM')
 			end,
 			[8] = function()
@@ -327,7 +336,7 @@ elseif L1UI.TBC then
 				PluginInstallFrame.Desc1:SetText('Please click the button below to apply Luckyones profile for Details! Damage Meter.')
 				PluginInstallFrame.Desc2:SetText('Importance: |cff4beb2cOptional|r')
 				PluginInstallFrame.Option1:Show()
-				PluginInstallFrame.Option1:SetScript('OnClick', function() L1UI:AddonSetupDT() end)
+				PluginInstallFrame.Option1:SetScript('OnClick', function() L1UI:Setup_Details() end)
 				PluginInstallFrame.Option1:SetText('Setup Details')
 			end,
 			[9] = function()