+++ ElvUI_LuckyoneUI/Profiles/AddOnSkins.lua
@@ -1,7 +1,7 @@
 local L1UI, E, L, V, P, G = unpack(select(2, ...))
 
 -- AddOnSkins Profile
-function L1UI:GetASProfile()
+function L1UI:Get_AddOnSkins_Profile()
 
 	local AS = unpack(AddOnSkins)