+++ ElvUI_LuckyoneUI/Profiles/ShadowLight.lua
@@ -1,8 +1,10 @@
 local L1UI, E, L, V, P, G = unpack(select(2, ...))
 if not L1UI.Retail then return end
 
+local GetAddOnMetadata = GetAddOnMetadata
+
 -- Shadow & Light Profile
-function L1UI:GetSLEProfile()
+function L1UI:Get_ShadowAndLight_Profile()
 
 	-- Shadow & Light Version
 	local version = GetAddOnMetadata('ElvUI_SLE', 'Version')